requirements.txt
streamlit==1.31.1
torch>=2.0.0
transformers==4.37.2
Pillow==10.2.0
Werkzeug==3.0.1
pandas==2.2.0
